refactor: reorganize creator
